requirements.txt
Flask==3.0.0
pandas==2.1.3
scikit-learn==1.3.2
tensorflow-cpu==2.15.0
seaborn==0.13.0
matplotlib==3.8.2
joblib==1.3.2
openpyxl==3.1.2
gunicorn==21.2.0
numpy==1.26.2
